Add RGB and SD card command handlers with configuration options
