+++ include/core/cli-handlers/rgb-cli-handler.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "managers/settings_manager.h"
+
+extern FSettings G_Settings;
+
+void handle_rgb_mode(int argc, char **argv);
+void handle_setrgb(int argc, char **argv);
