+++ include/core/cli-handlers/sd-cli-handler.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void handle_sd_config(int argc, char **argv);
+void handle_sd_pins_mmc(int argc, char **argv);
+void handle_sd_pins_spi(int argc, char **argv);
+void handle_sd_save_config(int argc, char **argv);
